Yary

+++ SUPERMARKET/Person.cs
@@ -29,7 +29,7 @@ namespace SUPERMARKET
 
         //PROPIEDADES ABSTRACTAS
 
-        public abstract int GetRating {get;  }
+        public abstract double GetRating {get;  }
 
         // CONSTRUCTORES
